+++ packages/authorization/repository/mock.go
@@ -56,3 +56,8 @@ func (repo *MockDescription) GetUserByRefreshToken(RefreshToken string) (models.
 func (s *MockDescription) DeleteSession(userID int64, publicSessionID string) error {
 	return MockData.Expected.Error
 }
+
+func (s *MockDescription) RequestUserPasswordReset(email, remoteAddr, confirmationKey string) (models.User, error) {
+	var user models.User
+	return user, MockData.Expected.Error
+}